RS/requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
pytube==15.0.0
google-api-python-client==2.118.0
python-dotenv==1.0.1 